code style update, added one FlatLaf theme to test

+++ src/main/java/org/gophie/config/SystemUtility.java
@@ -19,15 +19,17 @@
 package org.gophie.config;
 
 public class SystemUtility {
-    public static String getFileSizeString(float byteSize){
+    public static String getFileSizeString(float byteSize) {
         String result = byteSize + " Byte";
 
-        if(byteSize > 1000){
-            result = String.format("%.2f",(float)(byteSize / 1000)) + " KB";
-        }if(byteSize > 1000000){
-            result = String.format("%.2f",(float)(byteSize / 1000000)) + " MB";
-        }if(byteSize > 1000000000){
-            result = String.format("%.2f",(float)(byteSize / 1000000000)) + " GB";
+        if (byteSize > 1000) {
+            result = String.format("%.2f", byteSize / 1000) + " KB";
+        }
+        if (byteSize > 1000000) {
+            result = String.format("%.2f", byteSize / 1000000) + " MB";
+        }
+        if (byteSize > 1000000000) {
+            result = String.format("%.2f", byteSize / 1000000000) + " GB";
         }
 
         return result;
